// src/client/models/RestCreate.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type RestCreate = {
    /**
     * スタッフID
     */
    staff_id: number;
    /**
     * 出勤日
     */
    work_date: string;
    /**
     * 休憩開始時間
     */
    start_time?: (string | null);
    /**
     * 休憩終了時間
     */
    end_time?: (string | null);
};